+++ semver-gradle-plugin/test/kotlin/com/javiersc/semver/gradle/plugin/VersionTest.kt
@@ -11,36 +11,36 @@ internal class VersionTest {
     @Test
     fun `last version in current branch`() {
         initialCommitAnd {
+            fun cache() = GitCache(gitDir = git.repository.directory)
+
             resolve("2 commit.txt").createNewFile()
             git.add().addFilepattern(".").call()
             git.commit().setMessage("2 commit").call()
             git.tag().setName("v1.0.0").call()
-            GitCache(this).lastVersionInCurrentBranch("v").shouldBe(Version("1.0.0"))
+            cache().lastVersionInCurrentBranch("v").shouldBe(Version("1.0.0"))
 
             git.tag().setName("v1.0.0-alpha.1").call()
-            GitCache(this).lastVersionInCurrentBranch("v").shouldBe(Version("1.0.0"))
+            cache().lastVersionInCurrentBranch("v").shouldBe(Version("1.0.0"))
 
             resolve("3 commit.txt").createNewFile()
             git.add().addFilepattern(".").call()
             git.commit().setMessage("3 commit").call()
             git.tag().setName("v1.0.0-alpha.2").call()
-            GitCache(this).lastVersionInCurrentBranch("v").shouldBe(Version("1.0.0-alpha.2"))
+            cache().lastVersionInCurrentBranch("v").shouldBe(Version("1.0.0-alpha.2"))
 
             resolve("4 commit.txt").createNewFile()
             git.add().addFilepattern(".").call()
             git.commit().setMessage("4 commit").call()
 
             var isWarningLastVersionIsNotHigherVersion = false
-            GitCache(this).lastVersionInCurrentBranch("v") {
-                isWarningLastVersionIsNotHigherVersion = it
-            }
+            cache().lastVersionInCurrentBranch("v") { isWarningLastVersionIsNotHigherVersion = it }
             isWarningLastVersionIsNotHigherVersion.shouldBeTrue()
 
             resolve("5 commit.txt").createNewFile()
             git.add().addFilepattern(".").call()
             git.commit().setMessage("5 commit").call()
             git.tag().setName("v2.0.0-rc.2").call()
-            GitCache(this).lastVersionInCurrentBranch("v").shouldBe(Version("2.0.0-rc.2"))
+            cache().lastVersionInCurrentBranch("v").shouldBe(Version("2.0.0-rc.2"))
         }
     }
 }