+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 
 ### Added
 
+- `gitDir` property to `SemverExtension`
 - `printSemver` task depends on `prepareKotlinIdeaImport` task
-- Add `commits: Provider<Commit>` to `SemverExtension`
-- Add `commitsMaxCount: Provider<Int>` to `SemverExtension`
-- Add `semver.commitsMaxCount` property
+- `commits: Provider<Commit>` to `SemverExtension`
+- `commitsMaxCount: Provider<Int>` to `SemverExtension`
+- `semver.commitsMaxCount` property
 
 ### Changed